# perf: Fix the tactic sorting in TrtllmGenBatchedGemmRunner::getValidConfigIndices