<!-- .github/pull_request_template.md -->

## 📌 Description

The tactic sorting in `TrtllmGenBatchedGemmRunner::getValidConfigIndices` may not be as expected because different criteria are applied in different sorting function calls. This PR fixes the issue by applying these criteria in the same sorting function call. Moreover, the number of CTAs is estimated using `getNumCtas` to align with the kernel launch.

The performance of NVFP4 MoE is benchmarked on B200 using the following commands:
```bash
for (( batch_size=1; batch_size<=1024; batch_size=batch_size*2 )); do
    python3 flashinfer_benchmark.py \
        --routine trtllm_fp4_block_scale_moe \
	--num_tokens ${batch_size} \
	--hidden_size 7168 \
	--intermediate_size 256 \
	--num_experts 256 \
	--top_k 8 \
	--n_group 8 \
	--topk_group 4 \
	--routed_scaling_factor 2.5 \
	--use_routing_bias \
	--routing_method deepseek_v3 \
	--use_shuffled_weight \
	--verbose \
	--generate_repro_command
done
```

| Batch Size | Original Median Time (ms) | Modified Median Time (ms) |
| - | - | - |
| 1 | 0.024 | 0.020 |
| 2 | 0.026 | 0.021 |
| 4 | 0.034 | 0.023 |
| 8 | 0.043 | 0.027 |
| 16 | 0.094 | 0.047 |
| 32 | 0.057 | 0.057 |
| 64 | 0.079 | 0.079 |
| 128 | 0.098 | 0.098 |
| 256 | 0.144 | 0.144 |
| 512 | 0.234 | 0.234 |
| 1024 | 0.411 | 0.411 |


## 🔍 Related Issues

<!-- Link any related issues here -->

## 🚀 Pull Request Checklist

Thank you for contributing to FlashInfer! Before we review your pull request, please make sure the following items are complete.

### ✅ Pre-commit Checks

- [ ] I have installed `pre-commit` by running `pip install pre-commit` (or used your preferred method).
- [ ] I have installed the hooks with `pre-commit install`.
- [ ] I have run the hooks manually with `pre-commit run --all-files` and fixed any reported issues.

> If you are unsure about how to set up `pre-commit`, see [the pre-commit documentation](https://pre-commit.com/).

## 🧪 Tests

- [ ] Tests have been added or updated as needed.
- [ ] All tests are passing (`unittest`, etc.).

## Reviewer Notes

<!-- Optional: anything you'd like reviewers to focus on, concerns, etc. -->