@jinyangyuan-nvidia 2025-09-01
Thanks @yzh119. The perf data have been updated in the description.